% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut3.R
\name{cut3}
\alias{cut3}
\title{Cut a Numeric Variable into Intervals}
\usage{
cut3(x, i, what = c("breaks", "groups", "n_by_group", "n_intervals",
  "width", "width_0", "width_min", "width_max"), labels = NULL,
  expand = TRUE, crop = FALSE, simplify = TRUE, closed = c("left",
  "right", "both"), open_end = FALSE, ordered = FALSE,
  brackets = c("(", "[", ")", "]"), sep = ",", center_fun = NULL,
  optim_fun = NULL, format_fun = formatC, ...)
}
\arguments{
\item{x}{numeric vector to classify into intervals}

\item{format_fun}{formatting function}

\item{...}{additional arguments passed to \code{format_fun}}
}
\value{
a factor variable with levels of the form \code{"[a,b]"} or formatted means (character strings) unless \code{onlycuts} is \code{TRUE} in which case a numeric vector is returned
}
\description{
This behaves by default the same as Frank Harrel's \code{Hmisc::cut2},
and the code is essentially taken straight from the original.
The difference is that the \code{digits} argument was renamed to \code{levels.digits}
and that there are 2 additonal arguments : \code{format_fun} and \code{...} .
\code{format_fun} supports choosing another formating function than \code{format}, and \code{...}
is passed to the formatting function.
}
\details{
The motivation that led to modify the function was to be able to use \code{format_metric}
to have more readable cuts in some cases.

The original \code{digits} argument was renamed not to conflict with the \code{digits} argument used by formatting functions.
}
\examples{
set.seed(1)
numbers  <- cumsum(abs(rnorm(100,100,100)))
numbers  <- cumsum(abs(rnorm(10,10,10)))
cut3(numbers, 10,"groups",format_fun = format_metric, digits=2)
}
