% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutr-package.R
\docType{package}
\name{cutr-package}
\alias{cutr}
\alias{cutr-package}
\title{cutr: Enhanced cut And Useful Related Functions}
\description{
Expands formatting options of \code{base::cut.default} and \code{Hmisc::cut2},
supports factors, adhoc formating, group optimization and much more.
}
\keyword{internal}
