% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_metric.R
\name{format_metric}
\alias{format_metric}
\title{format a numeric vector using metric system}
\usage{
format_metric(
  x,
  prefixes = c("", "k", "M", "G"),
  digits = 3,
  smart_signif = FALSE,
  symbol = TRUE,
  base_unit = "",
  ...
)
}
\arguments{
\item{x}{a numeric vector}

\item{prefixes}{a vector or prefixes ('','k', 'M', ...), lower case symbols ('','kilo', 'mega' ...) or powers of tens (0, 3, 6 ...).}

\item{digits}{integer used by signif or smart_signif}

\item{smart_signif}{logical, if TRUE smart_signif will be used instead of signif}

\item{symbol}{logical, if TRUE, full symbol will be used, else prefix will be used}

\item{base_unit}{character, e.g. 'g'}

\item{...}{additional arguments passed to formatC}
}
\value{
formated character
}
\description{
format a numeric vector using metric system
}
\examples{
format_metric(c(-2500,0.15,3,5e6),c(0,3))
format_metric(c(-2500,0.15,3,5e6),"kilo")
format_metric(c(-2500,0.15,3,5e6),base_unit="\U{20AC}")
}
