% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cuts.R
\name{get_cuts}
\alias{get_cuts}
\title{get cut points}
\usage{
get_cuts(x, i, what, expand, crop, closed = "left", open_end,
  optim_fun = NULL, width_fun = NULL)
}
\arguments{
\item{width_fun}{}
}
\description{
get cut points
}
