% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_explicit.R
\name{middle}
\alias{middle}
\title{Middle of range}
\usage{
middle(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
a double
}
\description{
For easy use in center_fun argument of cut3
}
\examples{
middle(c(10,12,20))
}
