% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distribution.R
\name{plot_density}
\alias{plot_density}
\title{Filled Density Plot}
\usage{
plot_density(data, x, col = "orange")
}
\arguments{
\item{data}{the dataset in which is stored the variable to plot}

\item{x}{an integer, indicating the index of the column you want to visualise the distribution}

\item{col}{A string, indicating the color of the plot. Default is 'orange'.}
}
\value{
A plot with a density function
}
\description{
Create a simple and beautiful distribution plot
}
