% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{plot_rank}
\alias{plot_rank}
\title{Plots for outliers}
\usage{
plot_rank(dataset, i)
}
\arguments{
\item{dataset}{The  dataframe (or tibble) in which is stored the variable}

\item{i}{The number or the name of the variable}
}
\value{
a plot that is just perfecttp visualise outliers
}
\description{
Creates a plot with all the observations of a variable in order of rank, to identify outliers
}
