% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_taylor_russell.R
\name{plot_taylor_russell}
\alias{plot_taylor_russell}
\title{Interactive Taylor-Russell visualisation}
\usage{
plot_taylor_russell()
}
\value{
a nice graph of your selection hit rate,
}
\description{
gives a nice graphical visualisation of hit rates
}
\details{
You need no aguments. Just enjoy using the sliders in manipulate!
}
