% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_tests.R
\name{rename_shapes}
\alias{rename_shapes}
\title{rename shapes dimensions}
\usage{
rename_shapes(dataset, start_col, version)
}
\arguments{
\item{dataset}{The dataset to rename}

\item{start_col}{The index  of the first column of the dataset that is a shapes dimension (e.g. directing for management)}

\item{version}{For now only available 'basic' and 'management'}
}
\value{
A database, with the right column names for shapes
}
\description{
Rename the column of shapes from data extracts addingright labels
}
