% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smart_signif.R
\name{smart_signif}
\alias{smart_signif}
\title{Minimum preferred significant digits}
\usage{
smart_signif(x, digits = 3L, ...)
}
\arguments{
\item{x}{numeric, length 2 or more}

\item{digits}{integer, number of preferred remaining significant digits}

\item{...}{unused, kept for compatibility reasons}
}
\value{
numeric vector
}
\description{
Minimum preferred significant digits
}
\details{
Facilitate reducing numbers to their least \emph{distinguishable}
significant digits, where "distinguishable" means
"between neighbors". This means that if reducing more digits would
cause two neighbors to reduce to the same number, then the
reduction cannot take place.
}
\examples{
set.seed(1)
x  <- cumsum(abs(rnorm(10,100,100)))
smart_signif(x, 1)
smart_signif(x, 2)
smart_signif(x, 3)
}
\seealso{
\href{https://stackoverflow.com/q/51616332/3358272}{Original question on StackOverflow}
}
