% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cuts.R
\name{get_cuts}
\alias{get_cuts}
\title{get cut points}
\usage{
get_cuts(x, i, what, expand, crop, closed = "left", open_end,
  optim_fun = NULL, width_fun = NULL)
}
\arguments{
\item{optim_fun}{optimization function (used if \code{what = "group"})}

\item{width_fun}{left boundary function (used if \code{what = "width"})}
}
\value{
a vector of cut points
}
\description{
mainly used by \code{cutf3} but can be used to compute cut point without assessing
intervals to data points
}
\seealso{
cutf3
}
