% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cuts.R
\name{get_cuts}
\alias{get_cuts}
\title{get cut points}
\usage{
get_cuts(
  x,
  i,
  what,
  expand,
  crop,
  closed = "left",
  open_end,
  optim_fun = NULL,
  width_fun = NULL
)
}
\arguments{
\item{x}{numeric vector to classify into intervals}

\item{i}{numeric, character, or list, main parameter depending on \code{what}}

\item{what}{character, choices can be abreviated}

\item{expand}{logical, if TRUE cuts are added if necessary to cover min and max values}

\item{crop}{logical, if TRUE intervals which go past the min or max values will be cropped}

\item{closed}{character, which side of the intervals should be closed}

\item{open_end}{include in last interval on open side the values which fall on the last cutpoint}

\item{optim_fun}{optimization function (used if \code{what = "group"})}

\item{width_fun}{left boundary function (used if \code{what = "width"})}
}
\value{
a vector of cut points
}
\description{
mainly used by \code{smart_cut} but can be used to compute cut point without assessing
intervals to data points
}
\seealso{
?smart_cut
}
